+++ core/src/test/java/com/atlassian/migration/datacenter/core/aws/db/DatabaseMigrationServiceIT.java
@@ -24,6 +24,7 @@ import com.atlassian.migration.datacenter.core.aws.db.restore.SsmPsqlDatabaseRes
 import com.atlassian.migration.datacenter.core.aws.infrastructure.AWSMigrationHelperDeploymentService;
 import com.atlassian.migration.datacenter.core.aws.infrastructure.RemoteInstanceCommandRunnerService;
 import com.atlassian.migration.datacenter.core.aws.ssm.SSMApi;
+import com.atlassian.migration.datacenter.core.db.DatabaseClientTools;
 import com.atlassian.migration.datacenter.core.db.DefaultDatabaseExtractorFactory;
 import com.atlassian.migration.datacenter.core.fs.DefaultFileSystemMigrationReportManager;
 import com.atlassian.migration.datacenter.core.fs.FileSystemMigrationReportManager;
@@ -85,6 +86,9 @@ class DatabaseMigrationServiceIT {
     @Mock(lenient = true)
     ApplicationConfiguration configuration;
 
+    @Mock(lenient = true)
+    DatabaseClientTools databaseClientTools;
+    
     FileSystemMigrationReportManager reportManager = new DefaultFileSystemMigrationReportManager();
 
     @Mock
@@ -133,7 +137,7 @@ class DatabaseMigrationServiceIT {
     @Test
     void testDatabaseMigration() throws ExecutionException, InterruptedException, InvalidMigrationStageError, S3SyncFileSystemDownloader.CannotLaunchCommandException {
         MigrationStageCallback archiveStageTransitionCallback = new DatabaseArchiveStageTransitionCallback(migrationService);
-        DatabaseArchivalService databaseArchivalService = new DatabaseArchivalService(new DefaultDatabaseExtractorFactory(configuration), archiveStageTransitionCallback);
+        DatabaseArchivalService databaseArchivalService = new DatabaseArchivalService(new DefaultDatabaseExtractorFactory(configuration, databaseClientTools), archiveStageTransitionCallback);
 
         MigrationStageCallback uploadStageTransitionCallback = new DatabaseUploadStageTransitionCallback(this.migrationService);
         DatabaseArtifactS3UploadService s3UploadService = new DatabaseArtifactS3UploadService(() -> s3client, uploadStageTransitionCallback, reportManager);
